perf(core): remove two per-step transcript walks and two transcript copies
Round 1 answered the per-step cost finding with allocation micro-optimizations
inside each pass — a counting JSON sink, a streamed SHA-256, a Cow in
comparable_output. The constants got smaller; the pass count did not change.
This removes passes.

Token walks: 4 per step -> 2.

- `compact` discarded the conversation estimate it had just computed on both of
  its `None` paths, so `run_compaction_pass` re-walked the whole transcript in
  the `else` arm to recover it — on every step that did NOT compact, i.e. the
  common case, and eagerly, so it also ran when `summarize_overflow` was off
  and the value was never read. `compact_measured` returns the count on every
  path; `compact` stays as a thin wrapper for the tests.
- `ReceiptLedger::emit_step_receipt` ran its own `estimate_conversation_tokens`
  over the same unmutated slice the driver had just estimated, one call frame
  below it, purely so the manifest and `StepUsage` would agree. The driver now
  passes its value in, so they agree structurally. Replay/reconstruction
  callers that genuinely have no estimate use
  `emit_step_receipt_estimating`.

The two that remain are load-bearing: compaction must measure before deciding,
and a step must know its input size.

Transcript copies: 2 per successful step -> 1.

`run_model_call` took `messages.to_vec()` only so the `FnMut` retry closure
could own something, then cloned that again per attempt into
`CompletionRequest`. The closure now borrows the transcript and builds the
request's copy directly — one copy per attempt, which `CompletionRequest`
owning its messages makes unavoidable, instead of two per step.

And `recent_call_records` called `.into_owned()` on the `Cow` #700 introduced,
deep-copying every tool result's output on the borrowed path — the common one.
`CallRecord::output` is now `Option<Cow<'_, ToolOutput>>`, so the loop window
compares bytes the transcript is already holding. This window is rebuilt every
step over the whole turn, so the copy was quadratic in steps across a long turn.

Guards, because a pass count that nothing asserts comes back:
`a_step_walks_the_transcript_to_estimate_it_at_most_twice` counts real walks
through a 3-step turn (6 now; it was 12) and
`the_receipt_and_the_usage_record_report_one_estimate_per_step` pins the
invariant the deleted walk was maintaining by hand. Both live in a new
`driver/tests/compute_passes.rs` whose module doc says why pass counts, not
wall clock, are the thing to assert here.

Structural, from the repo's file-size gate rather than by choice:
stella-context's recall-facing loaders move to their own `candidates.rs` and
retrieval's tests to `retrieval/tests.rs`, which puts store.rs back under its
baseline. `emit_step_receipt` crossed clippy's argument limit, so the served
call identity (role, provider, model) is now one `ServedBy` struct rather than
three positional strings a caller can transpose. driver.rs (+2) and
driver/tests.rs (+1) are re-baselined: a module declaration and a threaded
argument.

Not done, and named rather than left to be rediscovered: each tool result's
output is still JSON-serialized and SHA-256'd twice per step
(`snapshot_result_identities` over the whole transcript, then `decompose`'s
`block_id` + `content_digest`), and `decompose` still re-hashes blocks it
already saw. Removing those needs a transcript-epoch invalidation contract so a
cache can tell an append from a compaction rewrite or a summarizer splice —
real work with real regression risk in loop detection, not a comment fix.
